feat(plugin): adapter plugin -- ceresdb and redis-jdbc
