+++ configure/etc/conf/plugins.properties
@@ -45,6 +45,8 @@ datacap-plugin-greptimedb=plugin/datacap-plugin-greptimedb/pom.xml
 datacap-plugin-questdb=plugin/datacap-plugin-questdb/pom.xml
 datacap-plugin-h2-tcp=plugin/datacap-plugin-h2-tcp/pom.xml
 datacap-plugin-clickhouse-http=plugin/datacap-plugin-clickhouse-http/pom.xml
+datacap-plugin-ceresdb=plugin/datacap-plugin-ceresdb/pom.xml
+datacap-plugin-redis-jdbc=plugin/datacap-plugin-redis-jdbc/pom.xml
 datacap-convert-txt=convert/datacap-convert-txt/pom.xml
 datacap-convert-json=convert/datacap-convert-json/pom.xml
 datacap-convert-xml=convert/datacap-convert-xml/pom.xml
