--- plugin/datacap-http-ceresdb/src/main/java/io/edurt/datacap/plugin/http/ceresdb/CeresDBPluginModule.java
@@ -1,38 +1,0 @@
-package io.edurt.datacap.plugin.http.ceresdb;
-
-import com.google.inject.multibindings.Multibinder;
-import io.edurt.datacap.spi.AbstractPluginModule;
-import io.edurt.datacap.spi.PluginService;
-import io.edurt.datacap.spi.PluginModule;
-import io.edurt.datacap.spi.PluginType;
-
-public class CeresDBPluginModule
-        extends AbstractPluginModule
-        implements PluginModule
-{
-    @Override
-    public String getName()
-    {
-        return "CeresDB";
-    }
-
-    @Override
-    public PluginType getType()
-    {
-        return PluginType.HTTP;
-    }
-
-    @Override
-    public AbstractPluginModule get()
-    {
-        return this;
-    }
-
-    protected void configure()
-    {
-        Multibinder<String> module = Multibinder.newSetBinder(this.binder(), String.class);
-        module.addBinding().toInstance(this.getClass().getSimpleName());
-        Multibinder<PluginService> plugin = Multibinder.newSetBinder(this.binder(), PluginService.class);
-        plugin.addBinding().to(CeresDBPlugin.class);
-    }
-}
