--- plugin/datacap-http-ceresdb/src/main/java/io/edurt/datacap/plugin/http/ceresdb/CeresDBResponse.java
@@ -1,24 +1,0 @@
-package io.edurt.datacap.plugin.http.ceresdb;
-
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-import lombok.ToString;
-
-import java.util.List;
-import java.util.Map;
-
-@Data
-@ToString
-@NoArgsConstructor
-@AllArgsConstructor
-@JsonIgnoreProperties(ignoreUnknown = true)
-@SuppressFBWarnings(value = {"EI_EXPOSE_REP", "EI_EXPOSE_REP2"})
-public class CeresDBResponse
-{
-    private int code;
-    private String message;
-    private List<Map<String, String>> rows;
-}
