+++ plugin/datacap-plugin-ceresdb/pom.xml
@@ -2,27 +2,22 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0"
          xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <modelVersion>4.0.0</modelVersion>
     <parent>
-        <artifactId>datacap</artifactId>
         <groupId>io.edurt.datacap</groupId>
+        <artifactId>datacap</artifactId>
         <version>2024.4.0-SNAPSHOT</version>
         <relativePath>../../pom.xml</relativePath>
     </parent>
-    <modelVersion>4.0.0</modelVersion>
 
-    <artifactId>datacap-plugin-h2-tcp</artifactId>
-    <description>DataCap - Plugin - H2 - TCP</description>
+    <artifactId>datacap-plugin-ceresdb</artifactId>
+    <description>DataCap - Plugin - CeresDB</description>
 
     <dependencies>
         <dependency>
             <groupId>io.edurt.datacap</groupId>
             <artifactId>datacap-spi</artifactId>
             <scope>provided</scope>
-        </dependency>
-        <dependency>
-            <groupId>com.h2database</groupId>
-            <artifactId>h2</artifactId>
-            <version>${h2.version}</version>
         </dependency>
     </dependencies>
