+++ plugin/datacap-plugin-ceresdb/src/main/java/io/edurt/datacap/plugin/CeresDBPlugin.java
@@ -1,0 +1,5 @@
+package io.edurt.datacap.plugin;
+
+public class CeresDBPlugin
+        extends Plugin
+{}
