+++ plugin/datacap-plugin-redis-jdbc/src/main/java/io/edurt/datacap/plugin/RedisJdbcPlugin.java
@@ -1,0 +1,5 @@
+package io.edurt.datacap.plugin;
+
+public class RedisJdbcPlugin
+        extends Plugin
+{}
